All migrations to Godot 4.4
This includes the addition of UID files. These are the solution I was asking for for replacements to loading scripts/scenes/resources by path. Instead, you load them by UIDs now. The catch is, we need to commit UID files to version control, and we also need to move UID files alongside the actual scene/resource/whatever file for it to be tracked by the system. I'm sure the way they implemented this (naively) is just "find uid file, map uid to name, and find file by name in same directory as uid file"